+++ front-end/src/app/admin/register/page.tsx
@@ -2,9 +2,12 @@
 
 import { CustomButton } from "@/components/Button";
 import { Input } from "@/components/Input";
+
 import { List, ListItem, Stack, Typography } from "@mui/material";
+import { useRouter } from "next/navigation";
 
 const Register = () => {
+  const router = useRouter();
   return (
     <Stack
       alignItems={"center"}
@@ -18,36 +21,36 @@ const Register = () => {
       </Typography>
       <Stack gap={4} sx={{ width: "334px" }}>
         <Input
+          inputHandler={() => {
+            console.log();
+          }}
           label="Нэр"
           placeholder=""
           helperText=""
-          inputHandler={() => {
-            console.log("Asd");
-          }}
         />
         <Input
+          inputHandler={() => {
+            console.log();
+          }}
           placeholder=""
           label="Имэйл хаяг"
-          helperText="Зөв имэйл хаяг оруулна уу"
-          inputHandler={() => {
-            console.log("Asd");
-          }}
+          helperText=""
         />
         <Input
+          inputHandler={() => {
+            console.log();
+          }}
           label="Нууц үг"
           placeholder=""
-          helperText="Нууц үг ижил биш байна"
-          inputHandler={() => {
-            console.log("Asd");
-          }}
+          helperText=""
         />
         <Input
+          inputHandler={() => {
+            console.log();
+          }}
           label="Нууц үг давтах "
           placeholder=""
           helperText=""
-          inputHandler={() => {
-            console.log("Asd");
-          }}
         />
         <List disablePadding sx={{ listStyleType: "disc", marginLeft: "20px" }}>
           <ListItem
@@ -60,58 +63,16 @@ const Register = () => {
               color: "#71717A",
             }}
           >
-            Том үсэг орсон байх
-          </ListItem>
-          <ListItem
-            disablePadding
-            sx={{
-              display: "list-item",
-              fontSize: "12px",
-              fontWeight: "400",
-              lineHeight: "20px",
-              color: "#71717A",
-            }}
-          >
-            Жижиг үсэг орсон байх
-          </ListItem>
-          <ListItem
-            disablePadding
-            sx={{
-              display: "list-item",
-              fontSize: "12px",
-              fontWeight: "400",
-              lineHeight: "20px",
-              color: "#71717A",
-            }}
-          >
-            Тоо орсон байх
-          </ListItem>
-          <ListItem
-            disablePadding
-            sx={{
-              display: "list-item",
-              fontSize: "12px",
-              fontWeight: "400",
-              lineHeight: "20px",
-              color: "#71717A",
-            }}
-          >
-            Тэмдэгт орсон байх
+            Том, жижиг үсэг, тоо, тэмдэгт орсон байх
           </ListItem>
         </List>
         <CustomButton
-          text="Нэвтрэх"
+          text="Бүртгүүлэх"
           textColor="background.paper"
-          handleClick={() => console.log("Button clicked!")}
-          bgColor="secondary.main"
-          hoverColor="primary.main"
-          height="36px"
-          border="secondary.main"
-        />
-        <CustomButton
-          text="Нэвтрэх"
-          textColor="background.paper"
-          handleClick={() => console.log("Button clicked!")}
+          handleClick={() => {
+            console.log("Button clicked!");
+            router.push("/login");
+          }}
           bgColor="secondary.main"
           hoverColor="primary.main"
           height="36px"
