+++ front-end/src/app/layout.tsx
@@ -1,4 +1,6 @@
-import type { Metadata } from "next";
+"use client";
+
+import { usePathname } from "next/navigation";
 import localFont from "next/font/local";
 import "./globals.css";
 import { MuiProvider } from "@/provider/MuiProvider";
@@ -17,16 +19,14 @@ const geistMono = localFont({
   weight: "100 900",
 });
 
-export const metadata: Metadata = {
-  title: "E-Commerce",
-  description: "Generated by create next app",
-};
-
 export default function RootLayout({
   children,
 }: Readonly<{
   children: React.ReactNode;
 }>) {
+  const pathname = usePathname();
+  console.log(pathname, " pathname pathname pathname pathname");
+
   return (
     <html lang="en">
       <body className={`${geistSans.variable} ${geistMono.variable}`}>
@@ -38,11 +38,12 @@ export default function RootLayout({
               minHeight: "100vh",
             }}
           >
-            <Header />
+            {pathname.includes("admin") ? <></> : <Header />}
             <Box component="main" sx={{ flex: 1, mt: "95px" }}>
               {children}
             </Box>
-            <Footer />
+
+            {pathname.includes("admin") ? <></> : <Footer />}
           </Box>
         </MuiProvider>
       </body>
