+++ front-end/src/app/register/page.tsx
@@ -3,10 +3,8 @@
 import { CustomButton } from "@/components/Button";
 import { Input } from "@/components/Input";
 import { List, ListItem, Stack, Typography } from "@mui/material";
-import { useRouter } from "next/navigation";
 
 const Register = () => {
-  const router = useRouter();
   return (
     <Stack
       alignItems={"center"}
@@ -19,10 +17,38 @@ const Register = () => {
         Бүртгүүлэх
       </Typography>
       <Stack gap={4} sx={{ width: "334px" }}>
-        <Input label="Нэр" placeholder="" helperText="" />
-        <Input placeholder="" label="Имэйл хаяг" helperText="" />
-        <Input label="Нууц үг" placeholder="" helperText="" />
-        <Input label="Нууц үг давтах " placeholder="" helperText="" />
+        <Input
+          label="Нэр"
+          placeholder=""
+          helperText=""
+          inputHandler={() => {
+            console.log("Asd");
+          }}
+        />
+        <Input
+          placeholder=""
+          label="Имэйл хаяг"
+          helperText="Зөв имэйл хаяг оруулна уу"
+          inputHandler={() => {
+            console.log("Asd");
+          }}
+        />
+        <Input
+          label="Нууц үг"
+          placeholder=""
+          helperText="Нууц үг ижил биш байна"
+          inputHandler={() => {
+            console.log("Asd");
+          }}
+        />
+        <Input
+          label="Нууц үг давтах "
+          placeholder=""
+          helperText=""
+          inputHandler={() => {
+            console.log("Asd");
+          }}
+        />
         <List disablePadding sx={{ listStyleType: "disc", marginLeft: "20px" }}>
           <ListItem
             disablePadding
@@ -34,17 +60,58 @@ const Register = () => {
               color: "#71717A",
             }}
           >
-            Том, жижиг үсэг, тоо, тэмдэгт орсон байх
+            Том үсэг орсон байх
+          </ListItem>
+          <ListItem
+            disablePadding
+            sx={{
+              display: "list-item",
+              fontSize: "12px",
+              fontWeight: "400",
+              lineHeight: "20px",
+              color: "#71717A",
+            }}
+          >
+            Жижиг үсэг орсон байх
+          </ListItem>
+          <ListItem
+            disablePadding
+            sx={{
+              display: "list-item",
+              fontSize: "12px",
+              fontWeight: "400",
+              lineHeight: "20px",
+              color: "#71717A",
+            }}
+          >
+            Тоо орсон байх
+          </ListItem>
+          <ListItem
+            disablePadding
+            sx={{
+              display: "list-item",
+              fontSize: "12px",
+              fontWeight: "400",
+              lineHeight: "20px",
+              color: "#71717A",
+            }}
+          >
+            Тэмдэгт орсон байх
           </ListItem>
         </List>
         <CustomButton
-          text="Бүртгүүлэх"
-          href={"/login"}
+          text="Нэвтрэх"
           textColor="background.paper"
-          handleClick={() => {
-            console.log("Button clicked!");
-            router.push("/login");
-          }}
+          handleClick={() => console.log("Button clicked!")}
+          bgColor="secondary.main"
+          hoverColor="primary.main"
+          height="36px"
+          border="secondary.main"
+        />
+        <CustomButton
+          text="Нэвтрэх"
+          textColor="background.paper"
+          handleClick={() => console.log("Button clicked!")}
           bgColor="secondary.main"
           hoverColor="primary.main"
           height="36px"
